requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pillow==10.3.0
reportlab==4.2.0
pandas==2.2.2
python-dotenv==1.0.1
pymongo==4.7.2
motor==3.4.0
aiosmtplib==3.0.1
python-multipart==0.0.9
email-validator==2.1.1
qrcode[pil]==7.4.2
aiofiles==23.2.1
